Add dynamic data fetch. Fix few issue.

+++ app/src/main/java/it/unibo/yahm/client/SpotholeService.kt
@@ -3,7 +3,6 @@ package it.unibo.yahm.client
 import io.reactivex.rxjava3.core.Observable
 import it.unibo.yahm.client.entities.Evaluations
 import it.unibo.yahm.client.entities.Leg
-import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
@@ -16,12 +15,13 @@ interface SpotholeService {
 
     @GET("roads/evaluations/relative")
     fun loadEvaluationsFromUserPerspective(@Query("latitude") latitude: Double,
-                                        @Query("longitude") longitude: Double,
-                                        @Query("radius") radius: Double): Observable<List<Leg>>
+                                           @Query("longitude") longitude: Double,
+                                           @Query("radius") radius: Float
+    ): Observable<List<Leg>>
 
     @GET("roads/evaluations/")
     fun loadEvaluations(@Query("latitude") latitude: Double,
                                         @Query("longitude") longitude: Double,
-                                        @Query("radius") radius: Double): Observable<List<Leg>>
+                                        @Query("radius") radius: Float): Observable<List<Leg>>
 
 }
